third_party/WebKit/LayoutTests/webaudio/audiocontext-suspend-resume-expected.txt

Test suspend/resume for an AudioContext

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.state is "paused"
PASS context.suspend() did not throw exception.
PASS context.suspend() did not throw exception.
PASS p1 = context.resume() did not throw exception.
PASS [object Object] is an instance of function Promise() { [native code] }
PASS p2 = context.resume() did not throw exception.
PASS [object Object] is an instance of function Promise() { [native code] }
PASS Promises from resume are not equal.
PASS context.state is "paused"
PASS Context resumed correctly.
PASS Context resumed correctly.
PASS context.state is "released"
PASS context.suspend() threw exception InvalidStateError: Failed to execute 'suspend' on 'AudioContext': cannot suspend an AudioContext that has been released.
PASS resume() on a released context rejected as expected
PASS successfullyParsed is true

TEST COMPLETE
